==> users/25_k-means_NHANES.R <==
# ===============================================================================================================
# k-means clustering.
# Version 1
# Created on 08/29/2022 by Rie Sadohara
# ===============================================================================================================

# READY TO BE COPIED TO TUTORIAL. === COPIED ON 08/30/2022.

# Set your working directory as to the main directory.
  Session --> Set working direHctory --> Choose directory.

# Name your main directory for future use.  
  main_wd <- file.path(getwd())

# Import source code to run the analyses to follow.
  source("lib/specify_data_dir.R")
  source("lib/k-means.R")

# You can come back to the main directory by:
  setwd(main_wd) 
  
# ---------------------------------------------------------------------------------------------------------------
# Specify the directory where your data is.
  SpecifyDataDirectory(directory.name = "eg_data/NHANES/Laboratory_data/")

# Here, we will run k-means analysis for each of Nutrients and Food categories.
  
# ===============================================================================================================
# Nutrient data, processed for clustering analyses.
# ===============================================================================================================

# Your input data should be a data frame with variables with non-zero variance. 
  nut_kmeansinput <- read.table("males50s_QCtotal_d_glu_body_meta_demo_Nut_rv.txt", 
                                sep="\t", header=T)
  
# Ensure your input file has the correct number of rows and columns.
  dim(nut_kmeansinput)
  
# Scale your input file and name it as k-means_input.
  kmeans_input <- scale(nut_kmeansinput)

# Specify the directory (folder) to save the results. (create a folder named as this if not done so)
  res_dir_nut = "males50s_Nut_k-means"

# Specify the prefix of filenames to be saved. 
  res_prefix_nut = "males50s_Nut"

# Run elbow, silhouette, and gap methods to find an optimum K (number of clusters). 
# Do not alter the name of the input file: kmeans_input. This function below assumes that
# the input is named as "kmeans_input". 
# You can only run those three methods for K = 1 through (number of observations - 1). 
# The gap method output will be printed on the Console. The gap values are plotted in 
# xxx_gapmethod.pdf.
  ChooseK(out.dir= res_dir_nut, out.prefix= res_prefix_nut)

# Look at the three figures generated by the ChooseK function above. The elbow method did not give a 
# distinct peak, but the silhouette and gap methods gave a peak at K=2. 
  
# With specific K values in mind, perform k-means analysis with one specified K.
# Also, change the file name to be saved as a PDF. 
  OneK(myK= 2, out.dir= res_dir_nut, out.fn = "males50s_Nut_K2")

# Or try multiple Ks and print the biplots in one panel.
# Likewise, change the file name to be saved as a PDF as necessary. 
# This uses the factoextra and gridExtra packages.  
  MultipleK(myKs = c(2,3,4,5), out.dir = res_dir_nut, out.fn = "males50s_Nut_K2-5")

# ===============================================================================================================
# Food category data, processed for clustering analyses.
# ===============================================================================================================

# Your input data should be a data frame with variables with non-zero variance. 
  cat_kmeansinput <- read.table("males50s_QCtotal_d_glu_body_meta_demo_Cat_rv.txt", 
                                sep="\t", header=T)

# Ensure your input file has the correct number of rows and columns.
  dim(cat_kmeansinput)

# Scale your input file and name it as k-means_input.
  kmeans_input <- scale(cat_kmeansinput)

# Specify the directory (folder) to save the results.
  res_dir_cat = "males50s_Cat_k-means"

# Specify the prefix of filenames to be saved. 
  res_prefix_cat = "males50s_Cat"

# Run elbow, silhouette, and gap methods to find an optimum K (number of clusters). 
# Do not alter the name of the input file: kmeans_input. This function below assumes that
# the input is named as "kmeans_input". 
# You can only run those three methods for K = 1 through (number of observations - 1). 
# The gap method output will be printed on the Console. The gap values are plotted in 
# xxx_gapmethod.pdf.
  ChooseK(out.dir= res_dir_cat, out.prefix= res_prefix_cat)

# Look at the three figures generated by the ChooseK function above. The elbow and gap method did not give a 
# distinct peak, but the silhouette method gave a peak at K=8. 

# With specific K values in mind, perform k-means analysis with one specified K.
# Also, change the file name to be saved as a PDF. 
  OneK(myK= 8, out.dir= res_dir_cat, out.fn = "males50s_Cat_K8")

# Or try multiple Ks and print the biplots in one panel.
# Likewise, change the file name to be saved as a PDF as necessary. 
# This uses the factoextra and gridExtra packages.  
  MultipleK(myKs = c(3,4,5,6), out.dir = res_dir_cat, out.fn = "males50s_Cat_K3-6")
  
# ---------------------------------------------------------------------------------------------------------------
# Come back to the main directory.
  setwd(main_wd)
