==> users/40_ordination_NHANES_males60to79_Lv3_Euclidean_sortfood.R <==
# ===============================================================================================================
# Sort SEQNs (columnnames of food_raw) in numeric order before building phyfoods object.
# Use Euclidean distance and run adonis. 
# Add metadata (GLU_index) before making a phyloseq object.
# Explore Axes 3 and 4 and other options.. 
# factor .... typo "labels" --> replaced with "levels"!! 
# colors "turquoise2", "goldenrod3", "mediumvioletred" --> replaced with "steelblue3", "gold3", "hotpink"
# to be consistent with PCA.
# Run adonis with Euclidean distance instead of unifrac.
# Version 1 - messy version, need to make a clean one.
# Created on 10/28/2022 by Rie Sadohara
# ===============================================================================================================

# Set your working directory to the main directory.
  Session --> Set working directory --> Choose directory.
  setwd("~/GitHub/dietarry_patterns")

# Name your main directory for future use. 
  main_wd <- file.path(getwd())

# ---------------------------------------------------------------------------------------------------------------
# Install the BiocManager package necessary for installing the phyloseq package. 
  if (!require("BiocManager",    quietly = TRUE))install.packages("BiocManager")
# Install the phyloseq package if you have not done so.
  BiocManager::install("phyloseq")
  
# Install the devtools package necessary for installing the pairwiseAdonis package. 
  if (!require("devtools",    quietly = TRUE))install.packages("devtools")
# install pairwise adonis function from Github. (https://github.com/pmartinezarbizu/pairwiseAdonis)
  devtools::install_github("pmartinezarbizu/pairwiseAdonis/pairwiseAdonis")
  
# ---------------------------------------------------------------------------------------------------------------

# load the necessary packages and the source code.
  library(vegan)
  library(phyloseq)
  library(ggplot2)
  library(ggtree)
  library(SASxport)
  library(pairwiseAdonis)
  source("lib/specify_data_dir.R")
  source("lib/ordination.R")
  source("lib/ggplot2themes.R")

# Load the distinct 100 colors for use.   
  distinct100colors <- readRDS("~/GitHub/R_Toolbox/distinct100colors.rda")

# You can come back to the main directory by:
  setwd(main_wd)

# Set working directory.
  SpecifyDataDirectory("eg_data/NHANES/Laboratory_data/")

# ===============================================================================================================
# Generate a phyloseq object using food, taxonomy, tree, and sample (metadata) for ordination. 
# ===============================================================================================================
  
# Load the necessary files for creating a phyloseq object.  
  
# Food
  # Load food OTU table - this is our food OTU data
  food_raw <- read.delim("Foodtree/Food_D12_FC_cc_f_males60to79_red_Lv3.dhydrt.otu.txt", row.names=1)
    head(food_raw, 1)  # food_raw has SEQN in a different order (X86435 X86515 X86817 X86831 ...).

  # Sort the columns (SEQN)...
    # The taxonomy is at the end of food - the nth column.
    taxonomycolumn <- length(colnames(food_raw))
    
    # order the column names of food, except the last column which has taxonomy.
    sortedcolnames_order <- order(colnames(food_raw)[ 1 : taxonomycolumn-1 ])
    
    # Sort the SEQNs but do not touch the taxonomy column at the end. 
    food <- food_raw[, c(sortedcolnames_order, taxonomycolumn) ] 
    
  # Take a look at the food file.
  # The column name of "food" is SEQN preceded with an 'X'.
  food[1:8, 1:8]
  
  # Format the food file and create a otu_table called OTU.
  PrepFood(data=food)
    head(OTU,1) # same order as food - X86435 X86515 X86817 X86831 ....

# # Taxonomy (tax)
#   # Load taxonomy file generated by the MakeFoodTree function.   
#   tax <- read.delim("Foodtree/Food_D12_FC_cc_f_males60to79_red_Lv3.taxonomy.txt")
#   
#   # Format the tax file and create a taxonomy table called TAX.
#   PrepTax(data=tax)
  
# Sample
  # Load the demographics data.
  demog <- read.xport("../Raw_data/DEMO_I.XPT")  # demog has SEQN in small-large order.
  
  # Load a dataset that has the "GLU_index" information. 
  glu <- read.delim( file="QCtotal_d_glu_body_meta.txt", sep= "\t", header= T )
  
    head(glu$SEQN) # This also has SEQN in small-large order.
    
  # Take out only the SEQN and GLU_index.
  SEQN_GLU <- glu[, c("SEQN", "GLU_index")]
      head(SEQN_GLU$SEQN)     # This also has SEQN in small-large order.
    
  # Add GLU_index to metadata. 
  demog_glu <- merge(x=SEQN_GLU, y=demog, all.x=TRUE, by="SEQN")
    head(demog_glu[, c("SEQN", "GLU_index")])  # This also has SEQN in small-large order.
  
  # Now, it has GLU_index.
  head(demog_glu, 2)
  
  # Put 'X' in front of the SEQN and define it as rownames.    
  rownames(demog_glu) <- paste("X", demog_glu$SEQN, sep="") 
  
  # # Prep metadata for generating a phyloseq object.  
  # PrepMeta_NHANES(data= demog_glu)
  #   # View(PrepMeta_NHANES)
  #   head(SAMPLES)                    # This is not in the small-large order, but in the original order. 
  #   # head(phyloseq::sample_data(phyfoods)) # Run after generating "phyfoods" object. this is also in the original order.
  # 
# # Food tree
#   # Load foodtree file generated by the MakeFoodTree function.   
#   foodtree <- read_tree("Foodtree/Food_D12_FC_cc_f_males60to79_red_Lv3.nwk")
#   # It is OK to see a message that says:
#   # "Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#   # Also defined by 'tidytree'"
#   
#   # Format the food tree and save it as 'TREE'. 
#   PrepTree(data=foodtree)
#   # It is OK to see the same message as the previous line. 
  
# # ---------------------------------------------------------------------------------------------------------------
# # Make a phyloseq object with OTU, TAX, samples, and foodtree by using the phyloseq function.
#   phyfoods <- phyloseq(OTU, TAX, SAMPLES, TREE)
#   # It is OK to see the same message as the previous line. 
#   
# # Check your metadata
# # Show the sample names. Change n to adjust the number of rows to show.
#   head(sample_names(phyfoods), n=6)  
#   
# # Show metadata. 
#   head(sample_data(phyfoods), n=2)
#   
# # Show only the columns (variables) of metadata. 
#   sample_variables(phyfoods)
#   
# # Check the level 1 foods in your food tree 
#   L1s <- tax_table(phyfoods)[, "L1"]
#   as.vector(unique(L1s))
# 
# Change to the folder called "Ordination" in your "Ordination" folder.
  SpecifyDataDirectory(directory.name = "eg_data/NHANES/Laboratory_data/Ordination/")
# 
# # ===============================================================================================================
# # Use your phyloseq object and perform ordination - WEIGHTED unifrac distance
# # ===============================================================================================================
#   
# # Perform Principal Coordinate Analysis (PCoA) with WEIGHTED unifrac distance of your food data.
#   # This may take a few minutes depending on your data size.
#   # e.g. a large phyloseq object (7.9 MB) could take a few minutes.
#   ordinated_w <- phyloseq::ordinate(phyfoods, method="PCoA", distance="unifrac", weighted=TRUE) 
#      
#      # some checking.
#      head(ordinated_w["vectors"])
#    
# # Save the percent variance explained by the axes as a vector to use in plots.  
#   eigen_percent_w <- ordinated_w$values$Relative_eig
# 
#   head(eigen_percent_w)  
# # Save the percent variance explained as a txt file.
#   Eigen(eigen.input = eigen_percent_w, 
#         output.fn="Food_D12_FC_cc_f_males60to79_red_Lv3_ord_WEIGHTED_eigen.txt")
# 
#   
# # ===============================================================================================================
# # Plot your ordination results - WEIGHTED
# # ===============================================================================================================
#   
# # Merge the first n axes to the metadata and save it as a txt file. 
# # The merged dataframe, 'meta_usersdf', will be used for plotting.
#   MergeAxesAndMetadata_NHANES(ord.object= ordinated_w, number.of.axes= 10, meta.data= demog_glu, 
#                               output.fn= "Food_D12_FC_cc_f_males60to79_red_Lv3_ord_WEIGHTED_meta_users.txt")
#     # View(MergeAxesAndMetadata_NHANES)
#     
# # Load the output again for plotting.
#   loaded_glu_w <- read.table("Food_D12_FC_cc_f_males60to79_red_Lv3_ord_WEIGHTED_meta_users.txt",
#                              sep="\t", header=T)
#   
#   table(loaded_glu_w$GLU_index)
# 
#      head(loaded_glu_w, 4)
#      write.table(loaded_glu_w[, c("SEQN", "GLU_index", "Axis.1")], "clipboard", sep="\t")
#      head(loaded_glu_w[, c("SEQN", "GLU_index", "Axis.1")] )
#      
#      write.table( as.data.frame(ordinated_w['vectors'])[, 1], "clipboard", sep="\t", row.names = T)
#      write.table( allvecdf[, c(1, ncol(allvecdf))], "clipboard", sep="\t", row.names = T)
#      
#   
#      allvec <- ordinated_w['vectors']
#      allvecdf <- as.data.frame(allvec)
#      allvecdf$SEQN <- row.names(allvecdf)
#      head(allvecdf[, c("SEQN","vectors.Axis.1")])
#      # sort by SEQN so that I can compare it with loaded_glu_w. 
#      allvecdf_s <- allvecdf[order(allvecdf$SEQN), ]
#      head(allvecdf_s[, c("SEQN","vectors.Axis.1")])
#      
#      # Confirmed that correct individuals have their correct Axis values.
#      # So, the separation in the ordination plots has correct individuals and subset labels.
# 
# # Convert the GLU_index as a factor to plot it in order.
#   loaded_glu_w$GLU_index <- factor(loaded_glu_w$GLU_index, levels= c("Normal", "Prediabetic", "Diabetic"))
#      table(loaded_glu_w$GLU_index)
#      
# 
# # ---------------------------------------------------------------------------------------------------------------
# # Plot Axis 1 and Axis 2 to show the separation of samples colored by Groups as in the metadata.
#   p1_w <- ggplot(loaded_glu_w, aes(x=Axis.1, y=Axis.2, color=GLU_index)) +
#     geom_point(aes(color= GLU_index), size=3) + 
#     scale_color_manual( values= c("steelblue3", "gold3", "hotpink")) +
#     xlab( paste("Axis.1 (", paste(round(eigen_percent_w[1]*100, 1)), "%)", sep="") ) +
#     ylab( paste("Axis.2 (", paste(round(eigen_percent_w[2]*100, 1)), "%)", sep="") ) +
#     no_grid + space_axes + theme(aspect.ratio = 1)
#   p1_w
#   
# # Save p1 as a png. 
#   ggsave("Food_D12_FC_cc_f_males60to79_red_Lv3_ord_WEIGHTED_Axis12_p1.png", 
#          p1_w, device="png", width=7, height=5.5, unit="in", dpi=300)
#   
# # You can add ellipses at a desired confidence level; but with this 
# # example data, there are too few samples per user to draw them. 
#   ellipses_w <- p1_w + stat_ellipse(level=0.95) 
#   ellipses_w
#   
# # Save ellipses as a png. 
#   ggsave("Food_D12_FC_cc_f_males60to79_red_Lv3_ord_WEIGHTED_Axis12_ellipses.png", 
#          ellipses_w, device="png", width=7, height=5.5, unit="in", dpi=300)
# 
# # ---------------------------------------------------------------------------------------------------------------
# # Plot Axis 2 and Axis 3 to show the separation of samples colored by Groups as in the metadata.
#   p1_w <- ggplot(loaded_glu_w, aes(x=Axis.2, y=Axis.3, color=GLU_index)) +
#     geom_point(aes(color= GLU_index), size=3) + 
#     scale_color_manual( values= c("steelblue3", "gold3", "hotpink")) +
#     xlab( paste("Axis.2 (", paste(round(eigen_percent_w[2]*100, 1)), "%)", sep="") ) +
#     ylab( paste("Axis.3 (", paste(round(eigen_percent_w[3]*100, 1)), "%)", sep="") ) +
#     no_grid + space_axes + theme(aspect.ratio = 1)
#   p1_w
#   
#   # Save p1 as a png. 
#   ggsave("Food_D12_FC_cc_f_males60to79_red_Lv3_ord_WEIGHTED_Axis23_p1.png", 
#          p1_w, device="png", width=7, height=5.5, unit="in", dpi=300)
#   
#   # You can add ellipses at a desired confidence level; but with this 
#   # example data, there are too few samples per user to draw them. 
#   ellipses_w <- p1_w + stat_ellipse(level=0.95) 
#   ellipses_w
#   
#   # Save ellipses as a png. 
#   ggsave("Food_D12_FC_cc_f_males60to79_red_Lv3_ord_WEIGHTED_Axis23_ellipses.png", 
#          ellipses_w, device="png", width=7, height=5.5, unit="in", dpi=300)
#   
# # ---------------------------------------------------------------------------------------------------------------
# # Plot Axis 3 and Axis 4 to show the separation of samples colored by Groups as in the metadata.
#   p1_w <- ggplot(loaded_glu_w, aes(x=Axis.3, y=Axis.4, color=GLU_index)) +
#     geom_point(aes(color= GLU_index), size=3) + 
#     scale_color_manual( values= c("steelblue3", "gold3", "hotpink")) +
#     xlab( paste("Axis.3 (", paste(round(eigen_percent_w[3]*100, 1)), "%)", sep="") ) +
#     ylab( paste("Axis.4 (", paste(round(eigen_percent_w[4]*100, 1)), "%)", sep="") ) +
#     no_grid + space_axes + theme(aspect.ratio = 1)
#   p1_w
#   
#   # Save p1 as a png. 
#   ggsave("Food_D12_FC_cc_f_males60to79_red_Lv3_ord_WEIGHTED_Axis34_p1.png", 
#          p1_w, device="png", width=7, height=5.5, unit="in", dpi=300)
#   
#   # You can add ellipses at a desired confidence level; but with this 
#   # example data, there are too few samples per user to draw them. 
#   ellipses_w <- p1_w + stat_ellipse(level=0.95) 
#   ellipses_w
#   
#   # Save ellipses as a png. 
#   ggsave("Food_D12_FC_cc_f_males60to79_red_Lv3_ord_WEIGHTED_Axis34_ellipses.png", 
#          ellipses_w, device="png", width=7, height=5.5, unit="in", dpi=300)  
# 
# # ---------------------------------------------------------------------------------------------------------------
# # Plot Axis 2 and Axis 4 to show the separation of samples colored by Groups as in the metadata.
#   p1_w <- ggplot(loaded_glu_w, aes(x=Axis.2, y=Axis.4, color=GLU_index)) +
#     geom_point(aes(color= GLU_index), size=3) + 
#     scale_color_manual( values= c("steelblue3", "gold3", "hotpink")) +
#     xlab( paste("Axis.2 (", paste(round(eigen_percent_w[2]*100, 1)), "%)", sep="") ) +
#     ylab( paste("Axis.4 (", paste(round(eigen_percent_w[4]*100, 1)), "%)", sep="") ) +
#     no_grid + space_axes + theme(aspect.ratio = 1)
#   p1_w
#   
#   # Save p1 as a png. 
#   ggsave("Food_D12_FC_cc_f_males60to79_red_Lv3_ord_WEIGHTED_Axis24_p1.png", 
#          p1_w, device="png", width=7, height=5.5, unit="in", dpi=300)
#   
#   # You can add ellipses at a desired confidence level; but with this 
#   # example data, there are too few samples per user to draw them. 
#   ellipses_w <- p1_w + stat_ellipse(level=0.95) 
#   ellipses_w
#   
#   # Save ellipses as a png. 
#   ggsave("Food_D12_FC_cc_f_males60to79_red_Lv3_ord_WEIGHTED_Axis24_ellipses.png", 
#          ellipses_w, device="png", width=7, height=5.5, unit="in", dpi=300)  
  
# ---------------------------------------------------------------------------------------------------------------
# The GLU_index groups look different. Use beta-diversity and adonis tests to see
# if they are actually distinct from one another.

# Prepare food OTU table.
  # Remove the taxonomy column from 'food'.
  food2 <- food[, !colnames(food) == "taxonomy"] 
  
  # Transpose food so that rows will be the SEQN (for which a distance matrix is calculated)
  food3 <- as.data.frame(t(food2))
  head(colnames(food3))    # columns have foods
  head(rownames(food3))    # rows have individuals
  
  # Sort individuals (rows) in order. **IMPORTANT!** 
  food3_s <- food3[order(rownames(food3)), ]
  head(rownames(food3_s))       # rows have individuals.

# ---------------------------------------------------------------------------------------------------------------
# Generate an Euclidean distance matrix of the users.  
  dist_matrix_e <- dist(food3_s, method = "euclidean", diag = T)
  dist_matrix_e <- dist(food3, method = "euclidean", diag = T)
  
# Save it as a txt file.
  write.table(as.matrix(dist_matrix_e), "Food_D12_FC_cc_f_males60to79_red_Lv3_Euclidean_dis_2.txt", 
              sep="\t", row.names=F, quote=F)
  
# Generate a TRUE or FALSE vector for presence in rownames (individuals) in food3.
  TF <- rownames(demog_glu) %in% rownames(food3_s)
  head(TF)
  TF
  
# Pick up only those in food3 from demog_glu. 
  demog_glu_male60to79 <-  demog_glu[TF, ]
  head(rownames(demog_glu_male60to79))

# Ensure the distance matrix and the demog_glu_subset have the same sample order.
# If this is not true, the tests below will be incorrect.
  identical(colnames(as.matrix(dist_matrix_e)), rownames(demog_glu_male60to79))
  
# ---------------------------------------------------------------------------------------------------------------
# Dispersion test and plot
  # vegan::betadisper computes centeroids and distance of each datapoint from it. 
  dispr_e <- vegan::betadisper(d= dist_matrix_e, demog_glu_male60to79$GLU_index)
  
  # Can show the centroids and dispersion of each group. 
  plot(dispr_e)
  
  # Or show the distance to centroid of each datapoint.
  boxplot(dispr_e, xlab = "")
  
  # Use dispr to do a permutation test for homogeneity of multivariate dispersion (are the clusters relatively of the same size?)
  vegan::permutest(dispr_e)
  
# ---------------------------------------------------------------------------------------------------------------
# adonis test with food3_s (sorted OTU) as the response and GLU_index of the individuals as the explanatory variable.   
  adonis(formula = food3_s ~ demog_glu_male60to79$GLU_index, method = "euclidean", permutations = 5000)
  
# If adonis is significant, use pairwise.adonis function to compare subgroups pairwise. 
  pairwise.adonis(dist_matrix_e, demog_glu_male60to79$GLU_index, perm = 5000)

# ---------------------------------------------------------------------------------------------------------------
# Come back to the main directory.
  setwd(main_wd)
